update usage message conidtions

+++ rm_n.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-echo "usage: $0 <dir> <n>" 1>&2
+if [[ $# -ne 2 ]]; then
+  echo "Usage: $0 <dir> <n>" >&2
+  exit 1
+fi
 
 find $1 -size +$2c -type f -delete
